# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus];

export const AccountStatus = {
  available: "available",
  "in-use": "in-use",
  "cooling-down": "cooling-down",
  archived: "archived",
} as const;

export interface Account {
  id: number;
  email: string;
  status: AccountStatus;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  cooldownDurationHours?: number | null;
  /** @nullable */
  cooldownStartedAt?: string | null;
  /** @nullable */
  cooldownEndsAt?: string | null;
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AccountInput {
  email: string;
  password: string;
  notes?: string;
  tags?: string[];
  cooldownDurationHours?: number;
}

export interface AccountUpdate {
  email?: string;
  password?: string;
  notes?: string;
  tags?: string[];
  cooldownDurationHours?: number;
}

export type StatusUpdateStatus =
  (typeof StatusUpdateStatus)[keyof typeof StatusUpdateStatus];

export const StatusUpdateStatus = {
  available: "available",
  "in-use": "in-use",
  "cooling-down": "cooling-down",
  archived: "archived",
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
}

export interface CooldownInput {
  durationHours: number;
}

export interface AccountStats {
  total: number;
  available: number;
  inUse: number;
  coolingDown: number;
  archived: number;
  readySoon: number;
}

export interface AccountExport {
  email: string;
  password?: string;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  cooldownDurationHours?: number | null;
  status: string;
}

export interface AccountImportPayload {
  accounts: AccountExport[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  errors?: string[];
}

export type ListAccountsParams = {
  search?: string;
  status?: ListAccountsStatus;
  tag?: string;
  sort?: ListAccountsSort;
};

export type ListAccountsStatus =
  (typeof ListAccountsStatus)[keyof typeof ListAccountsStatus];

export const ListAccountsStatus = {
  available: "available",
  "in-use": "in-use",
  "cooling-down": "cooling-down",
  archived: "archived",
} as const;

export type ListAccountsSort =
  (typeof ListAccountsSort)[keyof typeof ListAccountsSort];

export const ListAccountsSort = {
  "recently-used": "recently-used",
  "ready-first": "ready-first",
  "cooldown-ending-soon": "cooldown-ending-soon",
  "created-newest": "created-newest",
} as const;