# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountStatus } from "./accountStatus";

export interface Account {
  id: number;
  email: string;
  status: AccountStatus;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  cooldownDurationHours?: number | null;
  /** @nullable */
  cooldownStartedAt?: Date | null;
  /** @nullable */
  cooldownEndsAt?: Date | null;
  /** @nullable */
  lastUsedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}