# lib/api-zod/src/generated/types/accountExport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountExport {
  email: string;
  password?: string;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  /** @nullable */
  cooldownDurationHours?: number | null;
  status: string;
}