# lib/api-zod/src/generated/types/accountImportPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountExport } from "./accountExport";

export interface AccountImportPayload {
  accounts: AccountExport[];
}