# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  email: string;
  password: string;
  notes?: string;
  tags?: string[];
  cooldownDurationHours?: number;
}