# lib/api-zod/src/generated/types/accountStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountStats {
  total: number;
  available: number;
  inUse: number;
  coolingDown: number;
  archived: number;
  readySoon: number;
}