# lib/api-zod/src/generated/types/importResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImportResult {
  imported: number;
  skipped: number;
  errors?: string[];
}