# lib/api-zod/src/generated/types/statusUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dev Account Manager API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatusUpdateStatus } from "./statusUpdateStatus";

export interface StatusUpdate {
  status: StatusUpdateStatus;
}